refactor(components): :recycle: change RecentPosts into PostsGallery
